move utils from index to utils file
this avoid installing side effects in test and thus breaking source maps
